modified to handle byte arrays instead of long arrays and added panama to bytebin

+++ lucene/sandbox/src/java/org/apache/lucene/sandbox/rabitq/SubspaceOutput.java
@@ -1,3 +1,3 @@
 package org.apache.lucene.sandbox.rabitq;
 
-public record SubspaceOutput(float[][] cp, float[] x0, long[][] repackedBinXP) {}
+public record SubspaceOutput(float[][] cp, float[] x0, byte[][] repackedBinXP) {}
